requirements.txt
django==4.2
djangorestframework==3.14
django-split-settings==1.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
flake8==4.0
web3==6.1.0
drf-spectacular==0.26.1
pytest-django==4.5.2
isort==5.12.0
pytest-mock==3.10.0